reorganized file system and used pickle to build a working save system.
