+++ source/player.py
@@ -1,0 +1,6 @@
+#TODO: Other than just holding a ship, figure out what other things a Player object needs to hold
+#TODO: Test this file please
+class Player:
+    def __init__(self, ship):
+        #TODO: Move the UI parts of Ship here.
+        self.ship = ship
